Generate sparkSql support pg and optimized jdbc code
