+++ linkis-public-enhancements/linkis-datasource/linkis-metadata-query/service/jdbc/src/main/java/org/apache/linkis/metadata/query/service/PostgresqlMetaService.java
@@ -18,6 +18,7 @@
 package org.apache.linkis.metadata.query.service;
 
 import org.apache.linkis.datasourcemanager.common.util.json.Json;
+import org.apache.linkis.metadata.query.common.domain.GenerateSqlInfo;
 import org.apache.linkis.metadata.query.common.domain.MetaColumnInfo;
 import org.apache.linkis.metadata.query.common.service.AbstractDbMetaService;
 import org.apache.linkis.metadata.query.common.service.MetadataConnection;
@@ -101,6 +102,15 @@ public class PostgresqlMetaService extends AbstractDbMetaService<SqlConnection> 
       return connection.getColumns(database, table);
     } catch (SQLException | ClassNotFoundException e) {
       throw new RuntimeException("Fail to get Sql columns(获取字段列表失败)", e);
+    }
+  }
+
+  @Override
+  public GenerateSqlInfo querySparkSql(SqlConnection connection, String database, String table) {
+    try {
+      return connection.getSparkSql(database, table);
+    } catch (Exception e) {
+      throw new RuntimeException("Fail to get spark ddl sql (获取sparkDdlSql失败)", e);
     }
   }
 }
