+++ linkis-public-enhancements/linkis-datasource/linkis-metadata-query/service/jdbc/src/main/java/org/apache/linkis/metadata/query/service/postgres/SqlConnection.java
@@ -18,13 +18,10 @@
 package org.apache.linkis.metadata.query.service.postgres;
 
 import org.apache.linkis.common.conf.CommonVars;
-import org.apache.linkis.metadata.query.common.domain.MetaColumnInfo;
+import org.apache.linkis.metadata.query.service.AbstractSqlConnection;
 
-import org.apache.commons.collections.MapUtils;
 import org.apache.logging.log4j.util.Strings;
 
-import java.io.Closeable;
-import java.io.IOException;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
@@ -34,7 +31,7 @@ import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class SqlConnection implements Closeable {
+public class SqlConnection extends AbstractSqlConnection {
   private static final Logger LOG = LoggerFactory.getLogger(SqlConnection.class);
 
   private static final CommonVars<String> SQL_DRIVER_CLASS =
@@ -42,10 +39,6 @@ public class SqlConnection implements Closeable {
 
   private static final CommonVars<String> SQL_CONNECT_URL =
       CommonVars.apply("wds.linkis.server.mdm.service.postgre.url", "jdbc:postgresql://%s:%s/%s");
-
-  private Connection conn;
-
-  private ConnectMessage connectMessage;
 
   public SqlConnection(
       String host,
@@ -55,14 +48,7 @@ public class SqlConnection implements Closeable {
       String database,
       Map<String, Object> extraParams)
       throws ClassNotFoundException, SQLException {
-    connectMessage = new ConnectMessage(host, port, username, password, extraParams);
-    if (Strings.isBlank(database)) {
-      database = "";
-    }
-    conn = getDBConnection(connectMessage, database);
-    // Try to create statement
-    Statement statement = conn.createStatement();
-    statement.close();
+    super(host, port, username, password, Strings.isBlank(database) ? "" : database, extraParams);
   }
 
   public List<String> getAllDatabases() throws SQLException {
@@ -99,126 +85,30 @@ public class SqlConnection implements Closeable {
     }
   }
 
-  public List<MetaColumnInfo> getColumns(String schemaname, String table)
-      throws SQLException, ClassNotFoundException {
-    List<MetaColumnInfo> columns = new ArrayList<>();
-    String columnSql = "SELECT * FROM " + schemaname + "." + table + " WHERE 1 = 2";
-    PreparedStatement ps = null;
-    ResultSet rs = null;
-    ResultSetMetaData meta;
-    try {
-      List<String> primaryKeys = getPrimaryKeys(table);
-      ps = conn.prepareStatement(columnSql);
-      rs = ps.executeQuery();
-      meta = rs.getMetaData();
-      int columnCount = meta.getColumnCount();
-      for (int i = 1; i < columnCount + 1; i++) {
-        MetaColumnInfo info = new MetaColumnInfo();
-        info.setIndex(i);
-        info.setName(meta.getColumnName(i));
-        info.setType(meta.getColumnTypeName(i));
-        if (primaryKeys.contains(meta.getColumnName(i))) {
-          info.setPrimaryKey(true);
-        }
-        columns.add(info);
-      }
-    } finally {
-      closeResource(null, ps, rs);
-    }
-    return columns;
-  }
-
-  /**
-   * Get primary keys // * @param connection connection
-   *
-   * @param table table name
-   * @return
-   * @throws SQLException
-   */
-  private List<String> getPrimaryKeys(String table) throws SQLException {
-    ResultSet rs = null;
-    List<String> primaryKeys = new ArrayList<>();
-    DatabaseMetaData dbMeta = conn.getMetaData();
-    rs = dbMeta.getPrimaryKeys(null, null, table);
-    while (rs.next()) {
-      primaryKeys.add(rs.getString("column_name"));
-    }
-    return primaryKeys;
-  }
-
-  /**
-   * close database resource
-   *
-   * @param connection connection
-   * @param statement statement
-   * @param resultSet result set
-   */
-  private void closeResource(Connection connection, Statement statement, ResultSet resultSet) {
-    try {
-      if (null != resultSet && !resultSet.isClosed()) {
-        resultSet.close();
-      }
-      if (null != statement && !statement.isClosed()) {
-        statement.close();
-      }
-      if (null != connection && !connection.isClosed()) {
-        connection.close();
-      }
-    } catch (SQLException e) {
-      LOG.warn("Fail to release resource [" + e.getMessage() + "]", e);
-    }
-  }
-
-  @Override
-  public void close() throws IOException {
-    closeResource(conn, null, null);
-  }
-
   /**
    * @param connectMessage
    * @param database
    * @return
    * @throws ClassNotFoundException
    */
-  private Connection getDBConnection(ConnectMessage connectMessage, String database)
+  public Connection getDBConnection(ConnectMessage connectMessage, String database)
       throws ClassNotFoundException, SQLException {
+    String extraParamString =
+        connectMessage.extraParams.entrySet().stream()
+            .map(e -> String.join("=", e.getKey(), String.valueOf(e.getValue())))
+            .collect(Collectors.joining("&"));
     Class.forName(SQL_DRIVER_CLASS.getValue());
     String url =
         String.format(
             SQL_CONNECT_URL.getValue(), connectMessage.host, connectMessage.port, database);
-    if (MapUtils.isNotEmpty(connectMessage.extraParams)) {
-      String extraParamString =
-          connectMessage.extraParams.entrySet().stream()
-              .map(e -> String.join("=", e.getKey(), String.valueOf(e.getValue())))
-              .collect(Collectors.joining("&"));
+    if (!connectMessage.extraParams.isEmpty()) {
       url += "?" + extraParamString;
     }
     return DriverManager.getConnection(url, connectMessage.username, connectMessage.password);
   }
 
-  /** Connect message */
-  private static class ConnectMessage {
-    private String host;
-
-    private Integer port;
-
-    private String username;
-
-    private String password;
-
-    private Map<String, Object> extraParams;
-
-    public ConnectMessage(
-        String host,
-        Integer port,
-        String username,
-        String password,
-        Map<String, Object> extraParams) {
-      this.host = host;
-      this.port = port;
-      this.username = username;
-      this.password = password;
-      this.extraParams = extraParams;
-    }
+  @Override
+  public String getSqlConnectUrl() {
+    return SQL_CONNECT_URL.getValue();
   }
 }
